requirements.txt
absl-py==2.1.0
ADEval==1.1.0
anykeystore==0.2
apes==0.0.1
apex==0.9.10.dev0
Brotli @ file:///croot/brotli-split_1714483155106/work
causal-conv1d==1.4.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.1
cryptacular==1.6.2
cycler==0.12.1
defusedxml==0.7.1
einops==0.8.0
faiss-gpu==1.7.2
fastprogress==1.0.3
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.53.1
FrEIA==0.2
fsspec==2024.2.0
fvcore==0.1.5.post20221221
geomloss==0.2.6
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
greenlet==3.0.3
grpcio==1.66.1
huggingface-hub==0.24.5
hupper==1.12.1
idna==3.4
imageio==2.34.2
imgaug==0.4.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
mamba-ssm==2.2.2
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.1.post1
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmdet==2.25.3
mpmath @ file:///croot/mpmath_1690848262763/work
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.3
numpy-hilbert-curve==1.0.1
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
PasteDeploy==3.1.0
pbkdf2==1.3
pillow==10.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
portalocker==2.10.1
protobuf==3.20.1
pycocotools==2.0.8
pyparsing==3.1.2
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzorder==0.0.2
regex==2024.7.24
repoze.sendmail==4.4.1
requests==2.28.1
requests-oauthlib==2.0.0
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.32
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==0.8.15.dev0
tokenizers==0.19.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transaction==4.0
transformers==4.44.0
translationstring==1.4
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3==1.26.13
velruse==1.1.1
venusian==3.1.0
WebOb==1.8.8
Werkzeug==3.0.4
WTForms==3.1.2
wtforms-recaptcha==0.3.2
yacs==0.1.8
zope.deprecation==5.0
zope.interface==7.0.3
zope.sqlalchemy==3.1
